requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
distro==1.9.0
Django==5.2.8
django-extensions==4.1
feedparser==6.0.12
frozenlist==1.8.0
groq==0.34.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
magic-filter==1.0.12
multidict==6.7.0
mysql-connector-python==9.5.0
openai==2.17.0
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.2.1
selenium==4.38.0
sgmllib3k==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.3.1
yarl==1.22.0
